[FLINK-33441] Move Union restore tests
- Moving from nodes.exec.testutil to nodes.exec.stream
- Related commit: f362dcc
